// src/app/core/models/SaveHistory.ts
export interface SaveHistory{
    customerAccountId: number;
    requestId: string;
    comment: string;
    risk: string;
    commentDate: string;
    commentBy: string;
    status: string;
    accountStatus: string;
    approvedBy: string;
    approverIp: string;
    email: string;
    lastUpdatedBy: string;
    reasonForRework: string;
  }